+++ Scene/PlayScene.hpp
@@ -67,7 +67,7 @@ public:
     Engine::Sprite *dangerIndicator;
     Player *player;
     std::list<std::tuple<int, int, int>> bulletData;
-    std::list<std::tuple<int, int, int, int>> noteData;
+    std::list<std::tuple<int, int, int, float>> noteData;
     std::map<int, int> keyState;
     static Engine::Point GetClientSize();
     explicit PlayScene() = default;
